refactor(#179/resources): util, config 파일 global로 이동
