+++ src/main/java/inha/gdgoc/domain/auth/service/AuthService.java
@@ -1,6 +1,6 @@
 package inha.gdgoc.domain.auth.service;
 
-import inha.gdgoc.config.jwt.TokenProvider;
+import inha.gdgoc.global.config.jwt.TokenProvider;
 import inha.gdgoc.domain.auth.dto.request.UserLoginRequest;
 import inha.gdgoc.domain.auth.dto.response.LoginResponse;
 import inha.gdgoc.domain.auth.enums.LoginType;
@@ -28,7 +28,7 @@ import java.time.Duration;
 import java.util.Map;
 import java.util.Optional;
 
-import static inha.gdgoc.util.EncryptUtil.encrypt;
+import static inha.gdgoc.global.util.EncryptUtil.encrypt;
 
 @Slf4j
 @Service
@@ -36,6 +36,7 @@ import static inha.gdgoc.util.EncryptUtil.encrypt;
 public class AuthService {
 
     private final RefreshTokenService refreshTokenService;
+
     @Value("${google.client-id}")
     private String clientId;
 
@@ -63,9 +64,9 @@ public class AuthService {
 
         HttpEntity<MultiValueMap<String, String>> tokenRequest = new HttpEntity<>(params, headers);
         ResponseEntity<Map> tokenResponse = restTemplate.postForEntity(
-                "https://oauth2.googleapis.com/token",
-                tokenRequest,
-                Map.class
+            "https://oauth2.googleapis.com/token",
+            tokenRequest,
+            Map.class
         );
 
         String googleAccessToken = (String) tokenResponse.getBody().get("access_token");
@@ -76,10 +77,10 @@ public class AuthService {
         HttpEntity<Void> userInfoRequest = new HttpEntity<>(userInfoHeaders);
 
         ResponseEntity<Map> userInfoResponse = restTemplate.exchange(
-                "https://www.googleapis.com/oauth2/v2/userinfo",
-                HttpMethod.GET,
-                userInfoRequest,
-                Map.class
+            "https://www.googleapis.com/oauth2/v2/userinfo",
+            HttpMethod.GET,
+            userInfoRequest,
+            Map.class
         );
 
         // 3. Google에서 가져온 이름, 이메일로 가입된 정보가 없으면 회원가입, 있으면 로그인
@@ -90,9 +91,9 @@ public class AuthService {
         Optional<User> foundUser = userRepository.findByEmail(email);
         if (foundUser.isEmpty()) {
             return Map.of(
-                    "exists", false,
-                    "email", email,
-                    "name", name
+                "exists", false,
+                "email", email,
+                "name", name
             );
         }
 
@@ -100,29 +101,30 @@ public class AuthService {
 
         String jwtAccessToken = tokenProvider.generateGoogleLoginToken(user, Duration.ofHours(1));
         String refreshToken = refreshTokenService.getOrCreateRefreshToken(user, Duration.ofDays(1),
-                LoginType.GOOGLE_LOGIN);
+            LoginType.GOOGLE_LOGIN);
 
         ResponseCookie refreshCookie = ResponseCookie.from("refresh_token", refreshToken)
-                .httpOnly(true)
-                .secure(true)
-                .sameSite("None")
-                .domain(".gdgocinha.com")
-                .path("/")
-                .maxAge(Duration.ofDays(1))
-                .build();
+            .httpOnly(true)
+            .secure(true)
+            .sameSite("None")
+            .domain(".gdgocinha.com")
+            .path("/")
+            .maxAge(Duration.ofDays(1))
+            .build();
 
         // Set-Cookie 헤더로 추가
         log.info("Response Cookie에 저장된 Refresh Token: {}", refreshCookie.toString());
         response.addHeader(HttpHeaders.SET_COOKIE, refreshCookie.toString());
 
         return Map.of(
-                "exists", true,
-                "access_token", jwtAccessToken
+            "exists", true,
+            "access_token", jwtAccessToken
         );
     }
 
-    public LoginResponse loginWithPassword(UserLoginRequest userLoginRequest, HttpServletResponse response)
-            throws NoSuchAlgorithmException, InvalidKeyException {
+    public LoginResponse loginWithPassword(UserLoginRequest userLoginRequest,
+        HttpServletResponse response)
+        throws NoSuchAlgorithmException, InvalidKeyException {
         Optional<User> user = userRepository.findByEmail(userLoginRequest.email());
         if (user.isEmpty()) {
             return new LoginResponse(false, null);
@@ -135,16 +137,17 @@ public class AuthService {
         }
 
         String accessToken = tokenProvider.generateSelfSignupToken(foundUser, Duration.ofHours(1));
-        String refreshToken = refreshTokenService.getOrCreateRefreshToken(foundUser, Duration.ofDays(1),
-                LoginType.SELF_SIGNUP);
+        String refreshToken = refreshTokenService.getOrCreateRefreshToken(foundUser,
+            Duration.ofDays(1),
+            LoginType.SELF_SIGNUP);
 
         ResponseCookie refreshCookie = ResponseCookie.from("refresh_token", refreshToken)
-                .httpOnly(true)
-                .secure(true)
-                .sameSite("None")
-                .path("/")
-                .maxAge(Duration.ofDays(1))
-                .build();
+            .httpOnly(true)
+            .secure(true)
+            .sameSite("None")
+            .path("/")
+            .maxAge(Duration.ofDays(1))
+            .build();
 
         log.info("Response Cookie에 저장된 Refresh Token: {}", refreshCookie.toString());
         response.addHeader(HttpHeaders.SET_COOKIE, refreshCookie.toString());
