+++ src/main/java/inha/gdgoc/global/config/s3/S3Config.java
@@ -1,4 +1,4 @@
-package inha.gdgoc.config;
+package inha.gdgoc.global.config.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
