+++ src/main/java/inha/gdgoc/global/util/CookieUtil.java
@@ -1,4 +1,4 @@
-package inha.gdgoc.util;
+package inha.gdgoc.global.util;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
